Updated site visible for everyone
Updated dashboard so its visible for everyone.
Updated event so its visible for everyone.
Updated event image on the top of the card.
Updated companyoverview so its visible for everyone.
Updated company so its visible for everyone.
Updated http://127.0.0.1:8000/ goes to dashboard.
Updated when nog logged in there is a login button top right.